+++ js-web/vue-vite/firebase.config.js
@@ -10,4 +10,17 @@ window.CONFIG_FIREBASE_CLIENT = {
   appId: '1:833305455707:web:d9e03ba941195ae1'
 }
 
-window.CONFIG_VAPID_KEY = 'BC7v60EyiuKaBf9XNm1HV0gMlKhDWNXT6yA_fQ9l8u5s9SD7g3Hj7J915w-A4Hl3RsPLXCUroh_PrE9QKuhrD0I' // also generated in firebase website, may not be used
+window.CONFIG_VAPID_KEY = 'BC7v60EyiuKaBf9XNm1HV0gMlKhDWNXT6yA_fQ9l8u5s9SD7g3Hj7J915w-A4Hl3RsPLXCUroh_PrE9QKuhrD0I'
+
+// // created from firebase web dashboard
+// window.CONFIG_FIREBASE_CLIENT = {
+//   apiKey: "AIzaSyDXu2P9CQpaealHvyU6x_9DBR6MkGhKT1c",
+//   authDomain: "mybot-live.firebaseapp.com",
+//   databaseURL: "https://mybot-live.firebaseio.com",
+//   projectId: "mybot-live",
+//   storageBucket: "mybot-live.appspot.com",
+//   messagingSenderId: "632318854053",
+//   appId: "1:632318854053:web:ea2e76093423bd731c5e4c"
+// }
+// // also generated from firebase web dashboard, may not be used
+// window.CONFIG_VAPID_KEY = 'BF1kpd01rGl1ArPpe-DdN7FvKeneALT9dON1q0w7lgNeSS8w3u6_RsKcx5Ya_knMkH4SSsgtX_tLXviweWsq4Bs'